Idle / Walking Transition
Using a boolean parameter (IsWalking) on the units animator controller, the unit is now able to transition from an idle animation to a walking animation and back again.

This is currently done via the unit code, and is updated when the character moves.

Will hopefully get around to splitting the animations code from the movement code.